+++ tests/test_base_model.py
@@ -1,9 +1,16 @@
+import pulumi
+
 from laktory.models import Table
 from laktory.models import Schema
+from pulumi_random import RandomString
+
+env = RandomString("env", length=3, upper=False, numeric=False, special=False)
+schema_name = RandomString("schema", length=5, upper=False, numeric=False, special=False)
+
 
 schema = Schema(
-    name="stocks",
-    catalog_name="default",
+    name="${var.env}.${var.schema_name}",
+    catalog_name="${var.env}",
     tables=[
         Table(
             name="AAPL",
@@ -32,7 +39,11 @@ schema = Schema(
             ],
         ),
     ],
-    _vars={"dynamic_column": "low"},
+    _vars={
+        "dynamic_column": "low",
+        "env": env.id,
+        "schema_name": schema_name.id,
+    },
 )
 
 
@@ -40,6 +51,8 @@ def test_inject_vars():
     d0 = schema.model_dump()
     d1 = schema.inject_vars(d0)
     assert d1["tables"][-1]["columns"][0]["name"] == "low"
+    assert isinstance(d1["name"], pulumi.Output)
+    assert isinstance(d1["catalog_name"], pulumi.Output)
 
 
 def test_pulumi_dump():